Refactor / fixes
