+++ src/App.styles.js
@@ -8,8 +8,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background: linear-gradient(#ffa59e, #e26650)
-  
+  background: linear-gradient(#ffa59e, #e26650);
 }
 
 code {
